Add json run-time dependency to application list
Apparently, it is best practice to list all run-time dependencies in the list of `applications`. By not listing the `json` dependency in the application list, it precludes others from using your library if they need to make a release. For example, I'm building a nerves project (Elixir embedded software) that depends on `huex`, but it is failing because the release tools do not bundle the `json` library because it is not listed in the application block. In short, both _active applications_ as well as _library applications_ should be included in the application list as best practice. See this blog post for more information: 
http://blog.plataformatec.com.br/2016/07/understanding-deps-and-applications-in-your-mixfile/